move PowerUp and PowerDown component to Dashboard folder
