--- app/components/TransferDialog/styles.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const Description = styled.div`
-  padding: 10px 0;
-`;
